set empty startline to default value
instead of excluding the violation, we should use a default value to
ensure that the violation get included in the analysis.

since it might be a serious violation, it's always better to include
the violation even if the annotated line is incorrect. however, the
offset should not be more that a few lines at most, i guess.